define ballclass
added function definitions and refactored main game file.
