+++ game.py
@@ -2,16 +2,7 @@
 import tkinter
 import time
 import random
-
-# Class Def/Declaration
-class Ball:
-    def __init__(self, canvas, color):
-        self.canvas = canvas
-        self.id = canvas.create_oval(10, 10, 25, 25, fill = color)
-        self.canvas.move(self.id, 245, 100)
-
-    def draw(self):
-        self.canvas.move(self.id, 0, -1)
+import ballclass
 
 # Initilization
 tk = tkinter.Tk()
@@ -23,7 +14,7 @@ canvas.pack()
 tk.update()
 
 # Object creation
-ball = Ball(canvas, 'red')
+ball = ballclass.Ball(canvas, 'red')
 
 #Animation
 while 1:
